Add object approach, bug gallery
