+++ src/render.js
@@ -2,44 +2,10 @@
 
 var PAGE_SIZE = 12;
 var pageNumber = 0;
+var renderedPhotos = [];
 
-var templateElement = document.getElementById('picture-template');
-var elementToClone;
-
-/* Проверка наличия шаблона */
-if ('content' in templateElement) {
-  elementToClone = templateElement.content.querySelector('.picture');
-} else {
-  elementToClone = templateElement.querySelector('.picture');
-}
-
-/* Запалнение шаблона информацией о фото и изображением для рендеринга */
-var getPictureElement = function(data, container) {
-  var element = elementToClone.cloneNode(true);
-  var imageInElement = element.querySelector('img');
-
-  element.querySelector('.picture-comments').textContent = data.comments;
-  element.querySelector('.picture-likes').textContent = data.likes;
-
-  var contentImage = new Image();
-
-  contentImage.onload = function(evt) {
-    imageInElement.src = evt.target.src;
-    imageInElement.width = 182;
-    imageInElement.height = 182;
-    imageInElement.alt = data.date;
-  };
-
-  contentImage.onerror = function() {
-    element.classList.add('picture-load-failure');
-  };
-
-  contentImage.src = data.url;
-
-  container.appendChild(element);
-
-  return element;
-};
+/* Конструктор изображений */
+var Photo = require('./photo.js');
 
 /* Проверка на достижения нижнего края контейнера */
 var isBottomReached = require('./utilities').isBottomReached;
@@ -54,8 +20,8 @@ var renderPictures = function(picturesGeted, containerGeted, page) {
 
   var container = document.createDocumentFragment();
 
-  picturesGeted.slice(start, end).forEach(function(picture) {
-    getPictureElement(picture, container);
+  picturesGeted.slice(start, end).forEach(function(picture, number) {
+    renderedPhotos.push(new Photo(picture, number, container));
   });
 
   containerGeted.appendChild(container);
@@ -64,7 +30,12 @@ var renderPictures = function(picturesGeted, containerGeted, page) {
 module.exports = function(picturesGeted, containerGeted, reset) {
   if (reset) {
     pageNumber = 0;
-    containerGeted.innerHTML = '';
+
+    renderedPhotos.forEach(function(picture) {
+      picture.remove();
+    });
+
+    renderedPhotos = [];
   }
 
   while (isBottomReached(containerGeted) && isNextPageAvaible(picturesGeted, pageNumber, PAGE_SIZE)) {
